requirements.txt
fastapi
uvicorn
python-dotenv

langchain-core
langchain-openai
langchain-community
langchain-text-splitters
langchain-pinecone

pinecone-client
sentence-transformers

azure-storage-blob
pypdf
python-docx
unstructured
